add back the old yolo_v2_class library on Windows since it is still used by some projects
